Support for custom document IDs (#25)
Increases compatibility with existing datasets, e.g. in some cases it has been beneficial to store the document ID both as the "key" and in the data properties.

This allows both overriding the document ID to use a field other than id, and to not crash when there is another id or similar in the data.