Refactoring tests and how you get the transaction
Still need to update the example in the README and consider if there should be some more advanced test also.

The main test case and the City model now also show how you can work with an instance method and transactions. Seems like you need to wrap the transactional function in order to make it work. The `@async_transactional` decorator expects the first argument to be `transaction` and not for example `self`.